requirements.txt
fastapi
uvicorn
openai
pdfplumber
python-docx
faiss-cpu
requests
pydantic
numpy
python-dotenv 